change search component structure
