+++ lib/presentation/search/search_page.dart
@@ -1,21 +1,14 @@
 import 'package:flutter/material.dart';
 
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import 'package:yumemi_flutter_repo_search/domain/error.dart';
-import 'package:yumemi_flutter_repo_search/presentation/search/search_state_notifier.dart';
-import 'package:yumemi_flutter_repo_search/presentation/search/widget/error/error_widget.dart';
-import 'package:yumemi_flutter_repo_search/presentation/search/widget/list_item_shimmer.dart';
-import 'package:yumemi_flutter_repo_search/presentation/search/widget/result_list_view.dart';
+import 'package:yumemi_flutter_repo_search/presentation/search/widget/result_view.dart';
 import 'package:yumemi_flutter_repo_search/presentation/search/widget/search_app_bar.dart';
 import 'package:yumemi_flutter_repo_search/presentation/search/widget/search_bar.dart';
 
-class SearchPage extends ConsumerWidget {
+class SearchPage extends StatelessWidget {
   const SearchPage({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final searchState = ref.watch(searchStateNotifierProvider);
+  Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () => FocusManager.instance.primaryFocus?.unfocus(),
       behavior: HitTestBehavior.opaque,
@@ -23,49 +16,12 @@ class SearchPage extends ConsumerWidget {
         resizeToAvoidBottomInset: false,
         appBar: const SearchAppBar(),
         body: Column(
-          children: <Widget>[
+          children: const <Widget>[
             //検索フォーム
-            const SearchBar(),
-            const Divider(),
+            SearchBar(),
+            Divider(),
             //検索結果のリスト表示部分
-            Expanded(
-              flex: 8,
-              child: searchState.when(
-                //初期状態
-                uninitialized: () => const EnterTextView(),
-                //ローディング中
-                loading: () => const ListItemShimmer(),
-                //成功
-                success: (repositories, query, page, hasNext) => ResultListView(
-                  repoItems: repositories,
-                  hasNext: hasNext,
-                ),
-                //初期リクエスト失敗
-                failure: (exception) {
-                  if (exception is NoTextException) {
-                    return const EnterTextView();
-                  } else if (exception is NoInternetException) {
-                    return const NetworkErrorView();
-                  } else {
-                    return const ErrorView();
-                  }
-                },
-                //結果が空
-                empty: () => const NoResultView(),
-                //追加ローディング
-                fetchMoreLoading: (repositories, query, page) => ResultListView(
-                  repoItems: repositories,
-                  hasNext: true,
-                ),
-                //追加ローディングが失敗した
-                fetchMoreFailure: (repositories, query, page, exception) =>
-                    ResultListView(
-                  repoItems: repositories,
-                  hasNext: true,
-                  hasNextFetchError: true,
-                ),
-              ),
-            )
+            Expanded(flex: 8, child: ResultView())
           ],
         ),
       ),
